Refactor test helpers out of production modules
