+++ web-client/src/ui/store.testing.ts
@@ -1,0 +1,17 @@
+import {
+    clearUiStoreClientEventBindings,
+    resetUiStoreCatalogTracking,
+    resetUiStoreRuntimeCleanup,
+    restoreUiStoreBaseState,
+    subscribeToCatalog,
+    subscribeToRuntime,
+} from "./store";
+
+export function resetUiStoreForTesting() {
+    clearUiStoreClientEventBindings();
+    resetUiStoreRuntimeCleanup();
+    resetUiStoreCatalogTracking();
+    restoreUiStoreBaseState();
+    subscribeToRuntime();
+    subscribeToCatalog();
+}
