+++ web-client/test/__mocks__/sqljs-wasm-url.ts
@@ -1,0 +1,2 @@
+const wasmUrl = "mock-sql-wasm.wasm";
+export default wasmUrl;
